vim: removed minimal config version
Earlier I added a minimal init.vim with only a few customizations and no
plugins. The reason I added this was because my vim setup used to be
kinda slow and it wasn't worth it for quick edits. But after switching
my vim config to Lua, the speed problem is no longer there. So I am
removing the minimal config version.